Ramyaa preperations
Fixed a bug where genAction was calling policy_template instead of policy. This change allows the genPolicy to be random before policy is set, but also to actually follow a non-random policy after the model is trained. This made genPolicyAction redundant so it was removed.

Added stuff to main to make it play 1000 matches and print out win rate. Called before and after training.

Minor bug fixes for human playable version.